requirements.txt
selenium
webdriver-manager
markdownify
google-generativeai
tenacity
streamlit